requirements.txt
torch>=2.0.0
torchvision>=0.15.0
# carla>=0.9.13  # Install manually: pip install carla-0.9.x.whl from CARLA releases
opencv-python>=4.8.0
captum>=0.6.0
shap>=0.42.0
lime>=0.2.0.1
streamlit>=1.25.0
huggingface_hub>=0.16.0
numpy<2.0.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
tqdm>=4.65.0
datasets>=2.14.0
